# Speed up length on ARM