Vectorised instructions on ARM are pretty uniform, so I moved the implementation into a self-contained C file. This way all our trickery to detect AVX does not obscure the code.

Benchmarks on my M2 machine:

```
All
  Pure
    ascii-small
      length
        cons
          Text:     OK
            1.07 μs ± 105 ns, 82% less than baseline
          LazyText: OK
            1.11 μs ± 104 ns, 81% less than baseline
    ascii
      length
        cons
          Text:     OK
            1.31 ms ±  66 μs, 75% less than baseline
          LazyText: OK
            1.35 ms ± 124 μs, 74% less than baseline
    english
      length
        cons
          Text:     OK
            60.7 μs ± 3.9 μs, 82% less than baseline
          LazyText: OK
            64.4 μs ± 4.0 μs, 81% less than baseline
    russian
      length
        cons
          Text:     OK
            190  ns ±  13 ns, 81% less than baseline
          LazyText: OK
            192  ns ±  13 ns, 81% less than baseline
    japanese
      length
        cons
          Text:     OK
            204  ns ±  13 ns, 81% less than baseline
          LazyText: OK
            206  ns ±  13 ns, 81% less than baseline
```