@Lysxia 2024-09-14
The netbsd build is broken for unrelated reasons